:construction: Added notifications for low quantity of medication

+++ src/main/java/com/atlaspharmacy/atlaspharmacy/pharmacy/domain/Pharmacy.java
@@ -8,15 +8,15 @@ import javax.persistence.Table;
 @Table(name = "pharamcies")
 public class Pharmacy {
     @Id
-    private int id;
+    private Long id;
     private String name;
     private String description;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -36,7 +36,7 @@ public class Pharmacy {
         this.description = description;
     }
     public Pharmacy(){}
-    public Pharmacy(int id, String name, String description) {
+    public Pharmacy(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
